+++ music-api/src/main/java/music/endpoint/AdminEndpoint.java
@@ -34,11 +34,11 @@ public class AdminEndpoint {
     }
 
     @PostMapping("/dbSync")
-    public SyncResult syncTracksToDb() throws ReadOnlyFileException, CannotReadException, TagException, InvalidAudioFrameException, IOException {
+    public SyncResult syncTracksToDb(@RequestParam(defaultValue = "false") boolean forceUpdates) throws ReadOnlyFileException, CannotReadException, TagException, InvalidAudioFrameException, IOException {
         logger.info("Begin database sync");
         SyncResult syncResult = new SyncResult(new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), new ArrayList<>());
         List<Track> files = metadataService.getTracks();
-        trackService.upsertTracks(files, syncResult);
+        trackService.upsertTracks(files, syncResult, forceUpdates);
         trackService.deleteOrphanedTracksMetadata(files, syncResult);
         logger.info("Finished database sync");
         return syncResult;