+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # changelog
 
+## 1.1-2019.10.30-SNAPSHOT
+* feature: allow updates to be forced during sync (though request param `forceUpdates`)
+* fix: `updateByLocation` correctly gets bitrate from POJO
+
 ## 1.1-2019.10.29-SNAPSHOT
 * retrieve audio file header information (like file length)